test_gen_output/requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0
joblib>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0
